+++ 1.20/Forge/src/main/java/fuzs/respawninganimals/data/ModEntityTypesTagProvider.java
@@ -3,7 +3,6 @@ package fuzs.respawninganimals.data;
 import fuzs.puzzleslib.api.data.v1.AbstractTagProvider;
 import fuzs.respawninganimals.init.ModRegistry;
 import net.minecraft.core.HolderLookup;
-import net.minecraft.world.entity.EntityType;
 import net.minecraftforge.data.event.GatherDataEvent;
 
 public class ModEntityTypesTagProvider extends AbstractTagProvider.EntityTypes {
@@ -14,6 +13,6 @@ public class ModEntityTypesTagProvider extends AbstractTagProvider.EntityTypes {
 
     @Override
     protected void addTags(HolderLookup.Provider provider) {
-        this.tag(ModRegistry.PERSISTENT_ANIMALS_ENTITY_TYPE_TAG).add(EntityType.BEE);
+        this.tag(ModRegistry.PERSISTENT_ANIMALS_ENTITY_TYPE_TAG);
     }
 }