+++ 1.20/Forge/src/main/java/fuzs/respawninganimals/data/ModLanguageProvider.java
@@ -1,6 +1,7 @@
 package fuzs.respawninganimals.data;
 
 import fuzs.puzzleslib.api.data.v1.AbstractLanguageProvider;
+import fuzs.respawninganimals.init.ModRegistry;
 import net.minecraftforge.data.event.GatherDataEvent;
 
 public class ModLanguageProvider extends AbstractLanguageProvider {
@@ -11,7 +12,9 @@ public class ModLanguageProvider extends AbstractLanguageProvider {
 
     @Override
     protected void addTranslations() {
-        this.add("gamerule.persistentAnimals", "Persistent animals");
-        this.add("gamerule.persistentAnimals.description", "Animals will stay in the world forever and can only be reproduced by breeding.");
+        this.add(ModRegistry.PERSISTENT_ANIMALS_GAME_RULE, "Persistent animals");
+        this.addGameRuleDescription(ModRegistry.PERSISTENT_ANIMALS_GAME_RULE, "Animals will stay in the world forever and can only be reproduced from breeding.");
+        this.add(ModRegistry.ANIMAL_MOB_CAP_GAME_RULE, "Animal mob cap");
+        this.addGameRuleDescription(ModRegistry.ANIMAL_MOB_CAP_GAME_RULE, "Constant to help determine when to stop spawning animals in a world.");
     }
 }